atman-open-images-eval/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.8.0
    # via fiftyone
aiohttp==3.8.1
    # via
    #   datasets
    #   fsspec
aiosignal==1.2.0
    # via aiohttp
aleph-alpha-client==1.5.0
    # via -r requirements.in
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
argcomplete==2.0.0
    # via
    #   fiftyone
    #   voxel51-eta
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
backports-cached-property==1.0.2
    # via strawberry-graphql
boto3==1.24.21
    # via fiftyone
botocore==1.27.21
    # via
    #   boto3
    #   s3transfer
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   pip-tools
    #   strawberry-graphql
cycler==0.11.0
    # via matplotlib
dacite==1.6.0
    # via fiftyone
datasets==2.3.2
    # via -r requirements.in
deprecated==1.2.13
    # via fiftyone
dill==0.3.5.1
    # via
    #   datasets
    #   multiprocess
    #   voxel51-eta
dnspython==2.2.1
    # via eventlet
eventlet==0.33.1
    # via fiftyone
fiftyone==0.16.5
    # via -r requirements.in
fiftyone-brain==0.8.2
    # via fiftyone
fiftyone-db==0.3.0
    # via fiftyone
filelock==3.7.1
    # via huggingface-hub
fonttools==4.33.3
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.5.0
    # via datasets
future==0.18.2
    # via
    #   fiftyone
    #   voxel51-eta
glob2==0.7
    # via voxel51-eta
graphql-core==3.1.7
    # via strawberry-graphql
greenlet==1.1.2
    # via eventlet
h11==0.12.0
    # via
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via universal-analytics-python3
huggingface-hub==0.8.1
    # via datasets
hypercorn==0.13.2
    # via fiftyone
hyperframe==6.0.1
    # via h2
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
imageio==2.19.3
    # via scikit-image
jinja2==3.1.2
    # via fiftyone
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.1.0
    # via scikit-learn
kaleido==0.2.1
    # via fiftyone
kiwisolver==1.4.3
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.2
    # via fiftyone
mongoengine==0.20.0
    # via fiftyone
motor==2.5.1
    # via fiftyone
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.13
    # via datasets
ndjson==0.3.1
    # via
    #   fiftyone
    #   voxel51-eta
networkx==2.8.4
    # via scikit-image
numpy==1.22.4
    # via
    #   datasets
    #   fiftyone
    #   fiftyone-brain
    #   imageio
    #   matplotlib
    #   opencv-python-headless
    #   pandas
    #   pyarrow
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   voxel51-eta
opencv-python-headless==4.6.0.66
    # via
    #   -r requirements.in
    #   fiftyone
    #   voxel51-eta
packaging==21.3
    # via
    #   datasets
    #   fiftyone
    #   huggingface-hub
    #   matplotlib
    #   scikit-image
    #   voxel51-eta
pandas==1.4.2
    # via
    #   datasets
    #   fiftyone
patool==1.12
    # via voxel51-eta
pep517==0.12.0
    # via pip-tools
pillow==9.1.1
    # via
    #   -r requirements.in
    #   fiftyone
    #   imageio
    #   matplotlib
    #   scikit-image
    #   voxel51-eta
pip-tools==6.6.2
    # via -r requirements.in
plotly==4.14.3
    # via fiftyone
pprintpp==0.4.0
    # via fiftyone
priority==2.0.0
    # via hypercorn
psutil==5.9.1
    # via fiftyone
pyarrow==8.0.0
    # via datasets
pygments==2.12.0
    # via strawberry-graphql
pymongo==3.12.3
    # via
    #   fiftyone
    #   mongoengine
    #   motor
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   botocore
    #   matplotlib
    #   pandas
    #   strawberry-graphql
    #   voxel51-eta
python-multipart==0.0.5
    # via strawberry-graphql
pytz==2022.1
    # via
    #   fiftyone
    #   pandas
    #   voxel51-eta
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pywavelets==1.3.0
    # via scikit-image
pyyaml==6.0
    # via
    #   fiftyone
    #   huggingface-hub
requests==2.28.0
    # via
    #   aleph-alpha-client
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
    #   voxel51-eta
responses==0.18.0
    # via datasets
retrying==1.3.3
    # via
    #   fiftyone
    #   plotly
    #   voxel51-eta
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.6.0
    # via boto3
scikit-image==0.19.3
    # via
    #   fiftyone
    #   voxel51-eta
scikit-learn==1.1.1
    # via
    #   fiftyone
    #   fiftyone-brain
scipy==1.8.1
    # via
    #   fiftyone-brain
    #   scikit-image
    #   scikit-learn
sentinel==0.3.0
    # via strawberry-graphql
six==1.16.0
    # via
    #   eventlet
    #   plotly
    #   python-dateutil
    #   python-multipart
    #   retrying
    #   voxel51-eta
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sortedcontainers==2.4.0
    # via voxel51-eta
sse-starlette==0.10.3
    # via fiftyone
sseclient-py==1.7.2
    # via fiftyone
starlette==0.16.0
    # via
    #   fiftyone
    #   sse-starlette
strawberry-graphql==0.96.0
    # via fiftyone
tabulate==0.8.10
    # via
    #   fiftyone
    #   voxel51-eta
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2022.5.4
    # via scikit-image
toml==0.10.2
    # via hypercorn
tomli==2.0.1
    # via pep517
tqdm==4.64.0
    # via
    #   datasets
    #   huggingface-hub
typing-extensions==4.2.0
    # via
    #   huggingface-hub
    #   strawberry-graphql
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via voxel51-eta
universal-analytics-python3==1.1.1
    # via fiftyone
urllib3==1.26.9
    # via
    #   botocore
    #   requests
    #   responses
    #   voxel51-eta
voxel51-eta==0.7.1
    # via fiftyone
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via deprecated
wsproto==1.1.0
    # via hypercorn
xmltodict==0.13.0
    # via fiftyone
xxhash==3.0.0
    # via datasets
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
